pyproject.toml
[project]
name = "dell-lookup"
version = "0.1.0"
description = "Quick Dell API lookups and scripts and stuff"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "typer>=0.9.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
]

[project.scripts]
delly = "dell_lookup.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
